+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### To Release
 
+- [editor:fix] `space` will work for `text-field%` now
+- [editor:fix] `alt+backspace` will work for `text-field%` now
+
 ### v1.3.0
 
 - [shortcut:fix] `o+backspace` and `space` on improper editor